+++ packages/orq-rc/src/orq_ai_sdk/models/streamrunagentop.py
@@ -2193,7 +2193,7 @@ class AgentToolInputRunTools(BaseModel):
         StreamRunAgentAgentToolInputRunAgentsSchema, pydantic.Field(alias="schema")
     ]
 
-    id: Optional[str] = "01KG7A8PFWV0F09G633KD2V0Y2"
+    id: Optional[str] = "01KG7C48NR4FWSHWYC9Q41WW25"
 
     description: Optional[str] = None